crashblog/requirements.txt

chardet==5.1.0
charset-normalizer==3.1.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-fontawesome-5==1.0.18
django-js-asset==2.0.0
django-social-share==2.3.0
django-taggit==4.0.0
fontawesomefree==6.4.0
gunicorn==20.1.0
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.28.2
secure-smtplib==0.1.1
sqlparse==0.4.2
sympy==1.11.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
whichcraft==0.6.1
whitenoise==6.4.0
yarg==0.1.9

